Manutenção final 2, agora vai.

+++ seqProg.X/main.c
@@ -370,7 +370,7 @@ void main(void)
             case 3:
                             if(auxPasso & 0x80)
                             {
-                                setT1( (auxPasso & 0x7F) * 1000 );
+                                setT1( (unsigned long int)(auxPasso & 0x7F) * 1000 );
                                 meAtuadores = 4;
                             }
                             else
